Rename synapso -> synapso_core
