+++ .github/workflows/coverage.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Run tests with coverage
         run: |
           export PYTHONPATH=$PYTHONPATH:$(pwd)/src
-          pytest --cov=src/synapso --cov-report=xml --no-cov-on-fail
+          pytest --cov=src/synapso_core --cov-report=xml --no-cov-on-fail
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
