+++ .vscode/launch.json
@@ -18,7 +18,7 @@
             "module": "pytest",
             "args": [
                 "--durations=10",
-                "--cov=src/synapso",
+                "--cov=src/synapso_core",
                 "--cov-report=xml",
                 "${workspaceFolder}/tests"
             ],
@@ -31,7 +31,7 @@
             "module": "pytest",
             "args": [
                 "--durations=10",
-                "--cov=src/synapso",
+                "--cov=src/synapso_core",
                 "--cov-report=xml",
                 "${file}"
             ],
