+++ .vscode/settings.json
@@ -3,12 +3,12 @@
     "python.analysis.typeCheckingMode": "basic",
     "python.analysis.autoImportCompletions": true,
     "python.analysis.extraPaths": [
-        "./synapso"
+        "./synapso_core"
     ],
     "python.defaultInterpreterPath": "${workspaceFolder}/.venv/bin/python",
     "cursorpyright.analysis.autoImportCompletions": true,
     "cursorpyright.analysis.extraPaths": [
-        "./synapso"
+        "./synapso_core"
     ],
     "cursorpyright.analysis.typeCheckingMode": "basic",
     "[python]": {
