+++ src/synapso_core/chunking/factory.py
