+++ src/synapso_core/chunking/interface.py
