+++ src/synapso_core/ingestor/document_ingestor.py
