+++ src/synapso_core/persistence/factory.py
