+++ src/synapso_core/persistence/interfaces/__init__.py
