+++ src/synapso_core/persistence/interfaces/base_store.py
