+++ src/synapso_core/persistence/interfaces/meta_store.py
