+++ src/synapso_core/persistence/interfaces/private_store.py
