+++ src/synapso_core/persistence/interfaces/vector_store.py
