+++ src/synapso_core/persistence/models/base.py
