+++ src/synapso_core/persistence/models/cortex.py
