+++ src/synapso_core/query_manager/query_config.py
