+++ src/synapso_core/query_manager/query_manager.py
