+++ src/synapso_core/summarizer/__init__.py
