+++ src/synapso_core/summarizer/factory.py
