+++ src/synapso_core/summarizer/interface.py
