+++ src/synapso_core/vectorizer/__init__.py
