+++ src/synapso_core/vectorizer/factory.py
