+++ src/synapso_core/vectorizer/implementations/sentence_transformer_embeddings.py
