+++ src/synapso_core/vectorizer/interface.py
